requirements.txt
gtfs-realtime-bindings==1.0.0
PyYAML==6.0.0
schema==0.7.8
homeassistant==2022.7.7
